reorganize importer test files
